backend: rename main app.py to main.py; cleanup
